refactor(implementation): refactor `fixed_point` return behavior
Currently `fixed_point` returns the last `p` it found, regardless of whether it is a valid point
or just because the maximum number of iteration is reached.

This refactor add a warn log if MAX_ITER is reached, and return None then.

The print code in `find_root.py` is also changed. If `p` is None, the LaTeX table is still printed
if `-l` is used.